fix: :bug: check if customer exists before delete

+++ server/src/modules/customers/services/DeleteCustomerService.ts
@@ -1,5 +1,6 @@
 import { inject, injectable } from 'tsyringe';
 
+import { AppError } from '@shared/errors/AppError';
 import { ICustomersRepository } from '../repositories/ICustomersRepository'
 
 interface IServiceProps {
@@ -16,6 +17,12 @@ export class DeleteCustomerService {
   public async execute({
     customerId,
   }: IServiceProps): Promise<string> {
+    const customerToDelete = await this.customersRepository.findById(customerId)
+
+    if (!customerToDelete) {
+      throw new AppError('Customer not found!', 404)
+    }
+
     const customerIdDeleted = await this.customersRepository.delete(customerId)
 
     return customerIdDeleted
